MCOS/requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

Django

django-model-utils

mysql-python

celery

requests

Pillow<3.5.0

eventlet
kazoo

django-role-permissions
PyMySQL==0.7.11
sqlalchemy==1.1.14
django-adminlte2
python-memcached>=1.56
python-keystoneclient==3.10.0
-e git+https://github.com/ntk148v/CAL.git#egg=calplus
